requirements.txt
Flask==2.1.0
pytest==7.0.1; python_version > "3.0"
gunicorn==20.1.0
google-cloud-bigquery-storage==2.13.1
google-cloud-logging==3.0.0
google-cloud-pubsub==2.9.0
google-cloud-bigquery==2.33.0
python-dotenv